add test coverage for training programme controller

+++ src/main/java/co/uk/bransby/equinetrainingtrackerapi/api/models/dto/SkillTrainingSessionDto.java
@@ -16,6 +16,6 @@ public class SkillTrainingSessionDto {
     private TrainingMethodDto trainingMethod;
     private TrainingEnvironmentDto environment;
     private ProgressCode progressCode;
-    private Long trainingTime;
+    private Integer trainingTime;
     private String notes;
 }
